fix: quick search
